Stworzenie abstrakcji
Zmieniono klasę 'Lokalizacja' na typ klasy abstrakcyjnej. Dzięki temu napisałem też abstrakcyjne metody - te, które są wykonywane przez klasę 'Budynek', 'Poziom' i 'Pomieszczenie'.
I również wciśnięta została mała poprawka z double na int przy typie zwracanej wartości przez metodę 'getLiczbaOkien()' w klasie 'Pomieszczenie'.

+++ src/main/java/pl/put/poznan/buildingInfo/logic/Pomieszczenie.java
@@ -233,7 +233,7 @@ public class Pomieszczenie extends Lokalizacja{
      *
      * @return the nubmer of Building's windows from the room (describes the total windows number of the room)
      */
-    public double getLiczbaOkien() {
+    public int getLiczbaOkien() {
         return listaOkien.size();
     }
